src/monkeybrain/tools/file_meta.py

from win32com import client 
from pathlib import Path
from typing import List, Dict

"""
index_list = list( (index, namespace.GetDetailsOf(None, index)) for index in range(321)  )
[(0, 'Name'), (1, 'Größe'), (2, 'Elementtyp'), (3, 'Änderungsdatum'), (4, 'Erstelldatum'), (5, 'Letzter Zugriff'), (6, 'Attribute'), (7, 'Offlinestatus'), (8, 'Verfügbarkeit'), (9, 'Erkannter Typ'), (10, 'Besitzer'), (11, 'Art'), (12, 'Aufnahmedatum'), (13, 'Mitwirkende Interpreten'), (14, 'Album'), (15, 'Jahr'), (16, 'Genre'), (17, 'Dirigenten'), (18, 'Markierungen'), (19, 'Bewertung'), (20, 'Autoren'), (21, 'Titel'), (22, 'Betreff'), (23, 'Kategorien'), (24, 'Kommentare'), (25, 'Copyright'), (26, 'Titelnummer'), (27, 'Länge'), (28, 
'Bitrate'), (29, 'Geschützt'), (30, 'Kameramodell'), (31, 'Abmessungen'), (32, 'Kamerahersteller'), (33, 'Firma'), (34, 'Dateibeschreibung'), (35, 'Masterschlüsselwörter'), (36, 'Masterschlüsselwörter'), (37, ''), (38, ''), (39, ''), (40, ''), (41, ''), (42, 'Programmname'), (43, 'Dauer'), (44, 'Ist online'), (45, 'Periodisch wiederkehrend'), (46, 'Ort'), (47, 'Adressen der optionalen Teilnehmer'), (48, 'Optionale Teilnehmer'), (49, 'Organisatoradresse'), (50, 'Organisatorname'), (51, 'Erinnerungszeit'), (52, 'Adressen der erforderlichen Teilnehmer'), (53, 'Erforderliche Teilnehmer'), (54, 'Ressourcen'), (55, 'Besprechungsstatus'), (56, 'Status frei/besetzt'), (57, 'Gesamtgröße'), (58, 'Kontoname'), (59, ''), (60, 'Aufgabenstatus'), (61, 'Computer'), (62, 'Jahrestag'), (63, 'Name des Assistenten'), (64, 'Telefonnummer des Assistenten'), (65, 'Geburtstag'), (66, 'Geschäftsadresse'), (67, 
'Ort (geschäftlich)'), (68, 'Land/Region (geschäftlich)'), (69, 'Postfach (geschäftlich)'), (70, 'Postleitzahl (geschäftlich)'), (71, 'Bundesland/Provinz (geschäftlich)'), (72, 'Straße (geschäftlich)'), (73, 'Fax (geschäftlich)'), (74, 'Homepage (geschäftlich)'), (75, 'Rufnummer (geschäftlich)'), (76, 'Rückrufnummer'), (77, 'Autotelefon'), (78, 'Kinder'), (79, 'Zentrale Firmenrufnummer'), (80, 'Abteilung'), (81, 'E-Mail-Adresse'), (82, 'E-Mail2'), (83, 'E-Mail3'), (84, 'E-Mail-Liste'), (85, 'E-Mail-Anzeigename'), (86, 'Speichern unter'), (87, 'Vorname'), (88, 'Vollständiger Name'), (89, 'Geschlecht'), (90, 'Gegebener Name'), (91, 'Hobbies'), (92, 'Privatadresse'), (93, 'Ort (privat)'), (94, 'Land/Region (privat)'), (95, 'Postfach (privat)'), (96, 'Postleitzahl (privat)'), (97, 'Bundesland/Provinz (privat)'), (98, 'Straße (privat)'), (99, 'Fax (privat)'), (100, 'Rufnummer (privat)'), (101, 'Adressen für Chats'), (102, 'Initialen'), (103, 'Position'), (104, 'Bezeichnung'), (105, 'Nachname'), (106, 'Adresse'), (107, 'Zweiter Vorname'), (108, 'Mobiltelefon'), (109, 'Spitzname'), (110, 'Bürostandort'), (111, 'Weitere Adresse'), (112, 'Andere Stadt'), (113, 'Anderes Land/Region'), (114, 'Anderes Postfach'), (115, 'Andere Postleitzahl'), (116, 'Anderes Bundesland oder Provinz'), (117, 'Andere Straße'), (118, 'Pager'), (119, 'Persönlicher Titel'), (120, 'Stadt'), (121, 'Land/Region'), (122, 'Postfach'), (123, 'Postleitzahl'), (124, 'Bundesland/Provinz'), (125, 'Straße'), (126, 'Primäre E-Mail'), (127, 'Primäre Telefonnummer'), (128, 'Beruf'), (129, 'Ehepartner/Partner'), (130, 'Suffix'), (131, 'TTY/TTD-Telefon'), (132, 'Telex'), (133, 'Webseite'), (134, 'Inhaltstatus'), (135, 'Inhaltstyp'), (136, 'Erfassungsdatum'), (137, 'Archivierungsdatum'), (138, 'Vollendungsdatum'), (139, 
'Gerätekategorie'), (140, 'Verbindung hergestellt'), (141, 'Erkennungsmethode'), (142, 'Anzeigename'), (143, 'Lokaler Computer'), (144, 'Hersteller'), (145, 'Modell'), (146, 'Gekoppelt'), (147, 'Klassifizierung'), (148, 'Status'), (149, 'Gerätestatus'), (150, 'Clientkennung'), (151, 'Mitwirkende'), (152, 'Inhalt erstellt'), (153, 'Zuletzt gedruckt'), (154, 'Letzte Speicherung'), (155, 'Hauptabteilung'), (156, 'Dokument-ID'), (157, 'Seiten'), (158, 'Folien'), (159, 'Gesamtbearbeitungszeit'), (160, 'Wortanzahl'), (161, 'Fällig am'), (162, 'Enddatum'), (163, 'Dateianzahl'), (164, 'Dateierweiterung'), (165, 'Dateiname'), (166, 'Dateiversion'), (167, 'Kennzeichnungsfarbe'), (168, 'Kennzeichnungsstatus'), (169, 'Freier Speicherplatz'), (170, ''), (171, ''), (172, 'Gruppe'), (173, 'Freigabetyp'), (174, 'Bittiefe'), (175, 'Horizontale Auflösung'), (176, 'Breite'), (177, 'Vertikale Auflösung'), (178, 'Höhe'), (179, 'Wichtigkeit'), (180, 'Anlage?'), (181, 'Ist gelöscht'), (182, 'Verschlüsselungsstatus'), (183, 'Kennzeichnung vorhanden'), (184, 'Wurde beendet'), (185, 'Unvollständig'), (186, 'Lesestatus'), (187, 'Freigegeben'), (188, 'Ersteller'), (189, 'Datum'), (190, 'Ordnername'), (191, 'Ordnerpfad'), (192, 'Ordner'), (193, 'Teilnehmer'), (194, 'Pfad'), (195, 'Nach Ort'), (196, 'Typ'), (197, 'Kontaktnamen'), (198, 'Eintragstyp'), (199, 'Sprache'), (200, 'Letzter Besuch'), (201, 'Beschreibung'), (202, 'Verknüpfungsstatus'), (203, 'Verknüpfungsziel'), (204, 'URL'), (205, ''), (206, ''), (207, ''), (208, 'Medium erstellt'), (209, 'Veröffentlichungsdatum'), (210, 'Codiert durch'), (211, 'Folgennummer'), (212, 'Produzenten'), (213, 'Herausgeber'), (214, 'Staffelnummer'), (215, 'Untertitel'), (216, 'Benutzerweb-URL'), (217, 'Texter'), (218, ''), (219, 'Anlagen'), (220, 'BCC-Adressen'), (221, 'BCC'), (222, 'CC-Adressen'), (223, 'CC'), (224, 'Unterhaltungs-ID'), (225, 'Empfangsdatum'), (226, 'Absendungsdatum'), (227, 'Von Adressen'), (228, 'Von'), (229, 'Hat Anlagen'), (230, 'Absenderadresse'), (231, 'Absendername'), (232, 'Speicher'), (233, 'Empfängeradressen'), (234, 'Arbeitstitel'), (235, 'An'), (236, 'Laufzeit'), (237, 'Albuminterpret'), (238, 'Sortierung nach Albuminterpret'), (239, 'Album-ID'), (240, 'Sortierung nach Album'), (241, 'Sortierung nach mitwirkenden Interpreten'), (242, 'Beats pro Minute'), (243, 
'Komponisten'), (244, 'Sortierung nach Komponist'), (245, 'Disc'), (246, 'Ursprünglicher Schlüssel'), (247, 'Bestandteil einer Kompilation'), (248, 'Stimmung'), (249, 'Teil eines Satzes'), (250, 'Zeitraum'), (251, 'Farbe'), (252, 'Jugendschutz'), (253, 'Grund für Jugendschutzeinstufung'), (254, 'Verwendeter Speicherplatz'), (255, 'EXIF-Version'), (256, 'Ereignis'), (257, 'Lichtwert'), (258, 'Belichtungsprogramm'), (259, 'Belichtungszeit'), (260, 'Blendenzahl'), (261, 'Blitzlichtmodus'), (262, 'Brennweite'), (263, '35mm Brennweite'), (264, 'ISO-Filmempfindlichkeit'), (265, 'Objektivhersteller'), (266, 'Objektivmodell'), (267, 'Lichtquelle'), (268, 'Maximale Blende'), (269, 'Messmodus'), (270, 'Ausrichtung'), (271, 'Kontakte'), (272, 'Programmmodus'), (273, 'Sättigung'), (274, 'Abstand'), (275, 'Weißausgleich'), (276, 'Priorität'), (277, 'Projekt'), (278, 'Kanal'), (279, 'Folgenname'), (280, 'Untertitel (Closed Captions)'), (281, 'Wiederholung'), (282, 'Zweikanalton'), (283, 'Sendungsdatum'), (284, 'Sendungsbeschreibung'), (285, 'Aufnahmezeit'), (286, 'Senderrufzeichen'), (287, 'Fernsehsendername'), (288, 'Zusammenfassung'), (289, 'Schnipsel'), (290, 'Automatische Zusammenfassung'), (291, 'Relevanz'), (292, 'Dateibesitz'), (293, 'Sensitivität'), (294, 'Freigegeben für'), (295, 'Freigabestatus'), (296, ''), (297, 'Produktname'), (298, 'Produktversion'), (299, 'Supportlink'), (300, 'Quelle'), (301, 'Startdatum'), (302, 'Ist geteilt'), (303, 'Verfügbarkeitsstatus'), (304, 'Status'), (305, 'Abrechnungsinformationen'), (306, 'Abgeschlossen'), (307, 'Aufgabenbesitzer'), (308, 'Sortierung nach Titel'), (309, 'Gesamtdateigröße'), (310, 'Marken'), (311, 'Videokomprimierung'), (312, 'Regisseure'), (313, 'Datenrate'), (314, 'Bildhöhe'), (315, 'Einzelbildrate'), (316, 'Bildbreite'), 
(317, 'Kugelförmig'), (318, 'Stereo'), (319, 'Videoausrichtung'), (320, 'Gesamtbitrate')]
"""


# (298, 'Produktversion')

_lookup = {
    "Product version" : 298
}

from typing import Literal
key_types = Literal[
    "Product version"
]

def get_file_metadata(filepath:Path, metadata:List[key_types]) -> Dict[str, str]:
    """
    https://stackoverflow.com/a/63662404
    """

    path        = str(filepath.parent.absolute())
    filename    = str(filepath.name)

    dispatch = client.gencache.EnsureDispatch('Shell.Application', 0)

    namespace = dispatch.NameSpace(path)
    if namespace is None: return {}
    # Enumeration is necessary because ns.GetDetailsOf only accepts an integer as 2nd argument
    file_metadata = dict()
    item = namespace.ParseName(str(filename))
    index_list = list( ( _lookup[index_name], index_name) for index_name in metadata )
    for ind, attribute in index_list:
        attr_value = namespace.GetDetailsOf(item, ind)
        if attr_value:
            file_metadata[attribute] = attr_value
    
    return file_metadata