pyproject.toml
[project]
name = "monkeybrain"
version = "0.2.4"
description = "A CLI-Tool for launching TouchDesigner projects."
readme = "README.md"
license="MIT"
authors = [
    { name = "Wieland Hilker", email = "Wieland@PlusPlus.One" }
]
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "pywin32>=311",
]
keywords = ["TouchDesigner", "UV", "WindowsOnly"]


[project.urls]
Website = "https://plusplus.one"
Repository = "https://github.com/PlusPlusOneGmbH/Monkeybrain"


[tool.monkeybrain]
touchdesigner-version   = "2025.32050"
enforce-version         = "latest-version"
projectfile             = "Project.toe"

[tool.touchdesigner.TDPyEnvManagerContext]
mode    = "Python vEnv"
envName = ".venv"
installPath = "."
extraPaths  = ["from/project/toml"]

active          = true #can be omitted [?]
pythonVersion   = "3.11" #can be omitted [?]
autoSetup       = false #can be omitted [?]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[tool.hatch.build.targets.wheel]
packages = ["src/monkeybrain"]

[project.scripts]
"mb"   = "monkeybrain:entry"
